fasdfasdf

+++ dotnet.api/Models/OnibusModel.cs
@@ -13,7 +13,7 @@ namespace dotnet.api.Models
 
         public string Nome { get; set; }
 
-        public decimal Distancia { get; set; }
+        public double Distancia { get; set; }
 
         public bool Notifica { get; set; }
     }
